move Map.kt and PlotSurface.kt to coordplot root project folder
