+++ app/src/main/java/com/kodeco/android/coordplot/MainActivity.kt
@@ -4,7 +4,6 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import com.kodeco.android.coordplot.ui.theme.MyApplicationTheme
-import com.kodeco.android.coordplot.ui.theme.PlotSurface
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
