+++ app/src/main/java/com/kodeco/android/coordplot/PlotSurface.kt
@@ -1,4 +1,4 @@
-package com.kodeco.android.coordplot.ui.theme
+package com.kodeco.android.coordplot
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -14,6 +14,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.kodeco.android.coordplot.ui.theme.MyApplicationTheme
 
 @Composable
 fun PlotSurface() {
